+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_
 (which means it is unstable until 1.0).
 
 
+Unreleased
+~~~~~~~~~~
+Added
+-----
+- The middleware is now compatible with new-style (Django 1.10+) ``MIDDLEWARE``
+  as well as old-style ``MIDDLEWARE_CLASSES``.
+
+
 0.1.0 - 2016-05-08
 ~~~~~~~~~~~~~~~~~~